complete properties

+++ property/numerals.go
@@ -3,7 +3,7 @@ package property
 import "strings"
 
 type RomanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
@@ -23,38 +23,38 @@ var romanNumerals = []RomanNumeral{
 	{1, "I"},
 }
 
-func ConvertToRoman(n int) string {
+func ConvertToRoman(n uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range romanNumerals {
-		for n >= numeral.Value {
+		for n >= uint16(numeral.Value) {
 			result.WriteString(numeral.Symbol)
-			n -= numeral.Value
+			n -= uint16(numeral.Value)
 		}
 	}
 
 	return result.String()
 }
 
-func ConvertToRomanRec(n int) string {
+func ConvertToRomanRec(n uint16) string {
 	if n == 0 {
 		return ""
 	}
 
 	for _, numeral := range romanNumerals {
-		if n >= numeral.Value {
-			return numeral.Symbol + ConvertToRomanRec(n-numeral.Value)
+		if n >= uint16(numeral.Value) {
+			return numeral.Symbol + ConvertToRomanRec(n-uint16(numeral.Value))
 		}
 	}
 	return ""
 }
 
-func ConvertToArabic(roman string) int {
-	arabic := 0
+func ConvertToArabic(roman string) uint16 {
+	arabic := uint16(0)
 
 	for _, numeral := range romanNumerals {
 		for strings.HasPrefix(roman, numeral.Symbol) {
-			arabic += numeral.Value
+			arabic += uint16(numeral.Value)
 			roman = strings.TrimPrefix(roman, numeral.Symbol)
 
 		}
@@ -62,14 +62,14 @@ func ConvertToArabic(roman string) int {
 	return arabic
 }
 
-func ConvertToArabicRec(roman string) int {
+func ConvertToArabicRec(roman string) uint16 {
 	if roman == "" {
 		return 0
 	}
 
 	for _, numeral := range romanNumerals {
 		if strings.HasPrefix(roman, numeral.Symbol) {
-			return numeral.Value + ConvertToArabicRec(strings.TrimPrefix(roman, numeral.Symbol))
+			return uint16(numeral.Value) + ConvertToArabicRec(strings.TrimPrefix(roman, numeral.Symbol))
 		}
 	}
 	return 0
